feat: small terminal game for learning kana
To support learning Hiragana/Katakana symbols this game
will draw the kana at the top of the terminal "screen" and
the kana will float towards the bottom.

The player can remove them by typing the romaji form of the kana.

Currently only the basic kana (gajuon) are supported. More to come.