doc: have a plan
After the inital setup, let's have some ideas for how to expand this
little game, so that it is more useful for learning the Japanese kana
characters.

Topics like learning progress and the still crappy ui need some
polishing. And of course are we missing a lot of characters, yet.